Global fix bags

+++ Gym.Application/Members/Dto's/AdminStatsDto.cs
@@ -6,7 +6,7 @@ namespace Gym.Application.Members.Dto_s
 {
     public record AdminStatsDto
     {
-        public int CountActiveSubs { get; set; }
+        public long CountActiveSubs { get; set; }
         public decimal TotalProfitLastMonth { get; set; }
     }
 }
